+++ tower-http/src/compression/body.rs
@@ -1,5 +1,6 @@
 #![allow(unused_imports)]
 
+use crate::compression::CompressionLevel;
 use crate::{
     compression_utils::{AsyncReadBody, BodyIntoStream, DecorateAsyncRead, WrapBody},
     BoxError,
@@ -10,6 +11,9 @@ use async_compression::tokio::bufread::BrotliEncoder;
 use async_compression::tokio::bufread::GzipEncoder;
 #[cfg(feature = "compression-deflate")]
 use async_compression::tokio::bufread::ZlibEncoder;
+#[cfg(feature = "compression-zstd")]
+use async_compression::tokio::bufread::ZstdEncoder;
+
 use bytes::{Buf, Bytes};
 use futures_util::ready;
 use http::HeaderMap;
@@ -38,6 +42,19 @@ pin_project! {
     }
 }
 
+impl<B> Default for CompressionBody<B>
+where
+    B: Body + Default,
+{
+    fn default() -> Self {
+        Self {
+            inner: BodyInner::Identity {
+                inner: B::default(),
+            },
+        }
+    }
+}
+
 impl<B> CompressionBody<B>
 where
     B: Body,
@@ -55,6 +72,8 @@ where
             BodyInner::Deflate { inner } => inner.read.get_ref().get_ref().get_ref().get_ref(),
             #[cfg(feature = "compression-br")]
             BodyInner::Brotli { inner } => inner.read.get_ref().get_ref().get_ref().get_ref(),
+            #[cfg(feature = "compression-zstd")]
+            BodyInner::Zstd { inner } => inner.read.get_ref().get_ref().get_ref().get_ref(),
             BodyInner::Identity { inner } => inner,
         }
     }
@@ -68,6 +87,8 @@ where
             BodyInner::Deflate { inner } => inner.read.get_mut().get_mut().get_mut().get_mut(),
             #[cfg(feature = "compression-br")]
             BodyInner::Brotli { inner } => inner.read.get_mut().get_mut().get_mut().get_mut(),
+            #[cfg(feature = "compression-zstd")]
+            BodyInner::Zstd { inner } => inner.read.get_mut().get_mut().get_mut().get_mut(),
             BodyInner::Identity { inner } => inner,
         }
     }
@@ -93,6 +114,14 @@ where
                 .get_pin_mut(),
             #[cfg(feature = "compression-br")]
             BodyInnerProj::Brotli { inner } => inner
+                .project()
+                .read
+                .get_pin_mut()
+                .get_pin_mut()
+                .get_pin_mut()
+                .get_pin_mut(),
+            #[cfg(feature = "compression-zstd")]
+            BodyInnerProj::Zstd { inner } => inner
                 .project()
                 .read
                 .get_pin_mut()
@@ -127,6 +156,13 @@ where
                 .into_inner()
                 .into_inner()
                 .into_inner(),
+            #[cfg(feature = "compression-zstd")]
+            BodyInner::Zstd { inner } => inner
+                .read
+                .into_inner()
+                .into_inner()
+                .into_inner()
+                .into_inner(),
             BodyInner::Identity { inner } => inner,
         }
     }
@@ -140,6 +176,9 @@ type DeflateBody<B> = WrapBody<ZlibEncoder<B>>;
 
 #[cfg(feature = "compression-br")]
 type BrotliBody<B> = WrapBody<BrotliEncoder<B>>;
+
+#[cfg(feature = "compression-zstd")]
+type ZstdBody<B> = WrapBody<ZstdEncoder<B>>;
 
 pin_project_cfg! {
     #[project = BodyInnerProj]
@@ -162,6 +201,11 @@ pin_project_cfg! {
             #[pin]
             inner: BrotliBody<B>,
         },
+        #[cfg(feature = "compression-zstd")]
+        Zstd {
+            #[pin]
+            inner: ZstdBody<B>,
+        },
         Identity {
             #[pin]
             inner: B,
@@ -183,6 +227,11 @@ impl<B: Body> BodyInner<B> {
     #[cfg(feature = "compression-br")]
     pub(crate) fn brotli(inner: WrapBody<BrotliEncoder<B>>) -> Self {
         Self::Brotli { inner }
+    }
+
+    #[cfg(feature = "compression-zstd")]
+    pub(crate) fn zstd(inner: WrapBody<ZstdEncoder<B>>) -> Self {
+        Self::Zstd { inner }
     }
 
     pub(crate) fn identity(inner: B) -> Self {
@@ -209,6 +258,8 @@ where
             BodyInnerProj::Deflate { inner } => inner.poll_data(cx),
             #[cfg(feature = "compression-br")]
             BodyInnerProj::Brotli { inner } => inner.poll_data(cx),
+            #[cfg(feature = "compression-zstd")]
+            BodyInnerProj::Zstd { inner } => inner.poll_data(cx),
             BodyInnerProj::Identity { inner } => match ready!(inner.poll_data(cx)) {
                 Some(Ok(mut buf)) => {
                     let bytes = buf.copy_to_bytes(buf.remaining());
@@ -231,6 +282,8 @@ where
             BodyInnerProj::Deflate { inner } => inner.poll_trailers(cx),
             #[cfg(feature = "compression-br")]
             BodyInnerProj::Brotli { inner } => inner.poll_trailers(cx),
+            #[cfg(feature = "compression-zstd")]
+            BodyInnerProj::Zstd { inner } => inner.poll_trailers(cx),
             BodyInnerProj::Identity { inner } => inner.poll_trailers(cx).map_err(Into::into),
         }
     }
@@ -244,8 +297,8 @@ where
     type Input = AsyncReadBody<B>;
     type Output = GzipEncoder<Self::Input>;
 
-    fn apply(input: Self::Input) -> Self::Output {
-        GzipEncoder::new(input)
+    fn apply(input: Self::Input, quality: CompressionLevel) -> Self::Output {
+        GzipEncoder::with_quality(input, quality.into_async_compression())
     }
 
     fn get_pin_mut(pinned: Pin<&mut Self::Output>) -> Pin<&mut Self::Input> {
@@ -261,8 +314,8 @@ where
     type Input = AsyncReadBody<B>;
     type Output = ZlibEncoder<Self::Input>;
 
-    fn apply(input: Self::Input) -> Self::Output {
-        ZlibEncoder::new(input)
+    fn apply(input: Self::Input, quality: CompressionLevel) -> Self::Output {
+        ZlibEncoder::with_quality(input, quality.into_async_compression())
     }
 
     fn get_pin_mut(pinned: Pin<&mut Self::Output>) -> Pin<&mut Self::Input> {
@@ -278,8 +331,34 @@ where
     type Input = AsyncReadBody<B>;
     type Output = BrotliEncoder<Self::Input>;
 
-    fn apply(input: Self::Input) -> Self::Output {
-        BrotliEncoder::new(input)
+    fn apply(input: Self::Input, quality: CompressionLevel) -> Self::Output {
+        // The brotli crate used under the hood here has a default compression level of 11,
+        // which is the max for brotli. This causes extremely slow compression times, so we
+        // manually set a default of 4 here.
+        //
+        // This is the same default used by NGINX for on-the-fly brotli compression.
+        let level = match quality {
+            CompressionLevel::Default => async_compression::Level::Precise(4),
+            other => other.into_async_compression(),
+        };
+        BrotliEncoder::with_quality(input, level)
+    }
+
+    fn get_pin_mut(pinned: Pin<&mut Self::Output>) -> Pin<&mut Self::Input> {
+        pinned.get_pin_mut()
+    }
+}
+
+#[cfg(feature = "compression-zstd")]
+impl<B> DecorateAsyncRead for ZstdEncoder<B>
+where
+    B: Body,
+{
+    type Input = AsyncReadBody<B>;
+    type Output = ZstdEncoder<Self::Input>;
+
+    fn apply(input: Self::Input, quality: CompressionLevel) -> Self::Output {
+        ZstdEncoder::with_quality(input, quality.into_async_compression())
     }
 
     fn get_pin_mut(pinned: Pin<&mut Self::Output>) -> Pin<&mut Self::Input> {